+++ src/Sentry.Unity.Editor/Android/AndroidManifestConfiguration.cs
@@ -26,7 +26,7 @@ namespace Sentry.Unity.Editor.Android
 
         public AndroidManifestConfiguration()
             : this(
-                () => SentryScriptableObject.ConfiguredBuildtimeOptions(),
+                () => SentryScriptableObject.ConfiguredBuildTimeOptions(),
                 isDevelopmentBuild: EditorUserBuildSettings.development,
                 scriptingImplementation: PlayerSettings.GetScriptingBackend(BuildTargetGroup.Android))
         {
@@ -37,10 +37,10 @@ namespace Sentry.Unity.Editor.Android
             Func<(SentryUnityOptions?, SentryCliOptions?)> getOptions,
             bool isDevelopmentBuild,
             ScriptingImplementation scriptingImplementation,
-            IUnityLoggerInterceptor? interceptor = null)
+            ILogger? logger = null)
         {
             (_options, _sentryCliOptions) = getOptions();
-            _logger = _options?.DiagnosticLogger ?? new UnityLogger(_options ?? new SentryUnityOptions(), interceptor);
+            _logger = _options?.DiagnosticLogger ?? new UnityLogger(_options ?? new SentryUnityOptions(), logger);
 
             _isDevelopmentBuild = isDevelopmentBuild;
             _scriptingImplementation = scriptingImplementation;