+++ src/Sentry.Unity/SentryUnitySDK.cs
@@ -11,7 +11,7 @@ namespace Sentry.Unity
     {
         private readonly SentryUnityOptions _options;
         private IDisposable _dotnetSdk = null!;
-        private FileStream _lockFile = null!;
+        private FileStream? _lockFile;
 
         private SentryUnitySdk(SentryUnityOptions options)
         {
@@ -56,6 +56,13 @@ namespace Sentry.Unity
                         new ScreenshotAttachmentContent(options, SentryMonoBehaviour.Instance))));
             }
 
+            if (options.AttachViewHierarchy)
+            {
+                SentrySdk.ConfigureScope(s =>
+                    s.AddAttachment(new ViewHierarchyAttachment(
+                        new UnityViewHierarchyAttachmentContent(options, SentryMonoBehaviour.Instance))));
+            }
+
             if (options.NativeContextWriter is { } contextWriter)
             {
                 SentrySdk.ConfigureScope((scope) =>
@@ -96,7 +103,7 @@ namespace Sentry.Unity
             try
             {
                 // We don't really need to close, Windows would release the lock anyway, but let's be nice.
-                _lockFile.Close();
+                _lockFile?.Close();
             }
             catch (Exception ex)
             {