+++ test/Sentry.Unity.Tests/ScriptableSentryUnityOptionsTests.cs
@@ -20,7 +20,7 @@ namespace Sentry.Unity.Tests
                 persistentDataPath: "test/persistent/data/path");
         }
 
-        class TestOptionsConfiguration : ScriptableOptionsConfiguration
+        class TestOptionsConfiguration : SentryRuntimeOptionsConfiguration
         {
             public bool GotCalled;
             public override void Configure(SentryUnityOptions options) => GotCalled = true;
@@ -114,7 +114,7 @@ namespace Sentry.Unity.Tests
         {
             var optionsConfiguration = ScriptableObject.CreateInstance<TestOptionsConfiguration>();
             var scriptableOptions = ScriptableObject.CreateInstance<ScriptableSentryUnityOptions>();
-            scriptableOptions.OptionsConfiguration = optionsConfiguration;
+            scriptableOptions.RuntimeOptionsConfiguration = optionsConfiguration;
 
             scriptableOptions.ToSentryUnityOptions(isBuilding);