+++ test/Sentry.Unity.Tests/UnityLogHandlerIntegrationTests.cs
@@ -3,6 +3,7 @@ using System.Linq;
 using NUnit.Framework;
 using Sentry.Protocol;
 using Sentry.Unity.Integrations;
+using Sentry.Unity.Tests.SharedClasses;
 using Sentry.Unity.Tests.Stubs;
 using UnityEngine;
 
@@ -37,12 +38,11 @@ namespace Sentry.Unity.Tests
         }
 
         [Test]
-        public void CaptureLogFormat_LogStartsWithUnityLoggerPrefix_NotCaptured()
+        public void CaptureLogFormat_LogContainsSentryTag_NotCaptured()
         {
             var sut = _fixture.GetSut();
-            var message = $"{UnityLogger.LogPrefix}Test Message";
 
-            sut.CaptureLogFormat(LogType.Error, null, "{0}", message);
+            sut.CaptureLogFormat(LogType.Error, null, "{0}: {1}", UnityLogger.LogTag, "Test Message");
 
             Assert.AreEqual(0, _fixture.Hub.CapturedEvents.Count);
         }
@@ -229,6 +229,22 @@ namespace Sentry.Unity.Tests
             Assert.AreEqual(exception.GetType() + ": " + message, breadcrumb.Message);
             Assert.AreEqual("unity.logger", breadcrumb.Category);
             Assert.AreEqual(BreadcrumbLevel.Error, breadcrumb.Level);
+        }
+
+        [Test]
+        public void Register_RegisteredASecondTime_LogsWarningAndReturns()
+        {
+            var testLogger = new TestLogger();
+            _fixture.SentryOptions.DiagnosticLogger = testLogger;
+            _fixture.SentryOptions.Debug = true;
+            var sut = _fixture.GetSut();
+
+            // Edge-case of initializing the SDK twice with the same options object.
+            sut.Register(_fixture.Hub, _fixture.SentryOptions);
+
+            Assert.IsTrue(testLogger.Logs.Any(log =>
+                log.logLevel == SentryLevel.Warning &&
+                log.message.Contains("UnityLogHandlerIntegration has already been registered.")));
         }
     }
 }